refactor: move what a run reads into the writer package
The dispatch module was 920 lines against a 1000-line ceiling, and the
half of it that reads -- the metadata documents, the time selection, the
kernel directories and the clock kernel each of the run's clocks is
encoded with -- logs nothing and belongs with the rest of the code that
touches only files and SPICE.  It moves to spindoctor.cli.ck.inputs,
which leaves the driver holding the logging and the orchestration and
lets the reading be tested without a logger.

read_documents returns the files it could not read rather than logging
them, so the module stays log-free and the driver reports each by name.
