+++ src/spindoctor/cli/ck/__init__.py
@@ -65,6 +65,17 @@ from spindoctor.cli.ck.index import (
     build_ck_index,
     kernel_class_for_directory,
 )
+from spindoctor.cli.ck.inputs import (
+    Document,
+    clock_kernels,
+    furnish_supporting_kernels,
+    furnished,
+    kernel_paths,
+    read_documents,
+    recorded_basenames,
+    resolve_one,
+    select_by_time,
+)
 from spindoctor.cli.ck.kernel_file import first_data_record, write_ck_file
 from spindoctor.cli.ck.metakernel import build_meta_kernel_lines, write_meta_kernel
 from spindoctor.cli.ck.pointing import ImagePointing
@@ -93,6 +104,7 @@ __all__ = [
     'CkSegment',
     'CommentArea',
     'CoverageInterval',
+    'Document',
     'ImageEntry',
     'ImageFacts',
     'ImagePointing',
@@ -109,17 +121,25 @@ __all__ = [
     'build_meta_kernel_lines',
     'build_segment',
     'clock_is_defined',
+    'clock_kernels',
     'first_data_record',
+    'furnish_supporting_kernels',
+    'furnished',
     'group_for_output',
     'kernel_class_for_directory',
+    'kernel_paths',
     'output_basename',
     'read_comment_area',
+    'read_documents',
     'read_image_facts',
+    'recorded_basenames',
     'report_text',
     'reproduces_baseline',
     'reserved_comment_chars',
+    'resolve_one',
     'resolve_sclk_id',
     'rotation_angle_rad',
+    'select_by_time',
     'select_sclk_kernel',
     'utc_for_et',
     'write_ck_file',
